fix up dialog
