--- app/src/main/java/com/mjkj/listit/Composable/UI.kt
@@ -1,583 +1,0 @@
-package com.mjkj.listit.Composable
-
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import android.util.Log
-import android.widget.Toast
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material3.Button
-import androidx.compose.material3.Card
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ExposedDropdownMenuBox
-import androidx.compose.material3.ExposedDropdownMenuDefaults
-import androidx.compose.material3.FilledTonalButton
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalDrawerSheet
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.TextField
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
-import androidx.core.content.ContextCompat.startActivity
-import com.google.firebase.Firebase
-import com.google.firebase.auth.auth
-import com.google.firebase.firestore.firestore
-import com.mjkj.listit.Activity.ListActivity
-import com.mjkj.listit.Activity.MainActivity
-import com.mjkj.listit.Activity.ScrollListsActivity
-import com.mjkj.listit.Model.ListOfTasks
-import com.mjkj.listit.Model.User
-
-
-@Composable
-        /** This is a composable function that creates a button WITH a filled background and a label
-         *  this button is used as a main action button
-         *
-         * @param label: String - The text to be displayed on the button
-         * @param onClick: () -> Unit - The action to be performed when the button is clicked
-         * */
-fun ButtonFilled(label: String, onClick: () -> Unit) {
-    Button(onClick = { onClick() }) {
-        Text(label, fontSize = 25.sp)
-    }
-}
-
-@Composable
-        /** This is a composable function that creates a button WITHOUT a filled background and a with a passed label
-         *  this button is used as a secondary action button
-         *
-         * @param label: String - The text to be displayed on the button
-         * @param onClick: () -> Unit - The action to be performed when the button is clicked
-         **/
-fun ButtonTonalFilled(label: String, onClick: () -> Unit) {
-    FilledTonalButton(onClick = { onClick() }) {
-        Text(label, fontSize = 25.sp)
-    }
-}
-
-@Composable
-        /** This is a composable function that creates a text field with an outlined border
-         *
-         *  @param label: String - The text to be displayed on the label
-         *  @return text: String - The text entered into the text field
-         * */
-fun OutlinedTextField(label: String): String {
-    var text by remember { mutableStateOf("") }
-
-    androidx.compose.material3.OutlinedTextField(
-        value = text,
-        singleLine = true,
-        onValueChange = { text = it },
-        label = { Text(label) }
-    )
-    return text
-}
-
-@Composable
-        /**
-         * Composable function to display an app bar for a list screen.
-         *
-         * @param activity The title of the activity associated with the app bar.
-         */
-fun ListAppBar(
-    activity: String,
-    test: Activity,
-    listOfLists: List<List<String>>
-) {
-    val showDialog = remember {
-        mutableStateOf(false)
-    }
-
-    val showNavDrawer = remember {
-        mutableStateOf(false)
-    }
-
-    if ( (activity == "ListActivity" || activity == "ScrollListActivity")  && showDialog.value) {
-        Dialog(onDismissRequest = { showDialog.value = false },test)
-    }
-
-    if ((activity == "ListActivity" || activity == "ScrollListActivity") && showNavDrawer.value) {
-        NavDrawer(test, listOfLists = listOfLists)
-    }
-
-    Surface(
-        modifier = Modifier.fillMaxWidth(),
-        color = MaterialTheme.colorScheme.primary,
-        shape = RoundedCornerShape(bottomStart = 30.dp, bottomEnd = 30.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            IconButton(onClick = { showNavDrawer.value = changeState(showNavDrawer) }) {
-                Icon(
-                    imageVector = Icons.Filled.Menu,
-                    contentDescription = "Home",
-                    tint = Color.White
-                )
-            }
-            Text(
-                text = "List-it",
-                modifier = Modifier
-                    .weight(1f)
-                    .padding(horizontal = 16.dp),
-                textAlign = TextAlign.Center,
-                color = Color.White,
-                fontSize = 30.sp
-            )
-            IconButton(onClick = { showDialog.value = changeState(showDialog) }) {
-                Icon(
-                    imageVector = Icons.Filled.Add,
-                    contentDescription = "Add",
-                    tint = Color.White
-                )
-            }
-        }
-    }
-}
-
-fun changeState(state: MutableState<Boolean>): Boolean {
-    state.value = !state.value
-    return state.value
-}
-
-@Composable
-        /** This is a composable function that creates a dialog box
-         *
-         * @param onDismissRequest: () -> Unit - The action to be performed when the dialog is dismissed
-         * */
-fun Dialog(onDismissRequest: () -> Unit, parentActivity: Activity) {
-    var showCreateListContent by remember { mutableStateOf(false) }
-
-    Dialog(onDismissRequest = { onDismissRequest() }) {
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(650.dp)
-                .padding(16.dp),
-            shape = RoundedCornerShape(16.dp),
-        ) {
-            Column {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(25.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-
-                ) {
-                    TextButton(onClick = { showCreateListContent = false }) {
-                        Text(text = "Join List")
-                    }
-                    TextButton(onClick = { showCreateListContent = true }) {
-                        Text(text = "Create List")
-                    }
-                }
-                HorizontalDivider()
-                if(showCreateListContent){
-                    CreateListContent(parentActivity)
-
-                }else{
-                    JoinListContent(parentActivity)
-                }
-            }
-        }
-    }
-
-}
-
-@Composable
-        /** This is a composable function that joins to an existing a new list
-         *
-         * */
-
-fun JoinListContent(parentActivity: Activity){
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Spacer(modifier = Modifier.padding(5.dp))
-        Text(text = "Join an existing list", fontSize = 20.sp)
-        Spacer(modifier = Modifier.padding(25.dp))
-        var listCode: String = OutlinedTextField("List Code")
-        Spacer(modifier = Modifier.padding(20.dp))
-
-
-        Text(text = "Enter the code of the list you want to join", fontSize = 14.sp)
-        Spacer(modifier = Modifier.padding(20.dp))
-        HorizontalDivider(modifier = Modifier.height(5.dp))
-        Spacer(modifier = Modifier.padding(75.dp))
-        ButtonFilled("Join") {
-            Log.d("D", "JoinListCode: $listCode")
-            if(listCode.length != 6){
-                Toast.makeText(parentActivity, "Codes contain 6 symbols", Toast.LENGTH_SHORT).show()
-            }
-            else{
-                val db = Firebase.firestore
-                val user = User.createUser(Firebase.auth.currentUser?.uid,Firebase.auth.currentUser?.displayName.toString(),Firebase.auth.currentUser?.email.toString())
-                db.collection("lists").document(listCode).get()
-                    .addOnSuccessListener { document ->
-                        if (document != null) {
-                                db.collection("users").document(Firebase.auth.currentUser!!.uid)
-                                    .get()
-                                    .addOnSuccessListener { document ->
-                                        if (document != null) {
-                                            val currentList = document.data?.get("lists")
-                                            if (currentList == null) {
-                                                val tempCurrentList = mutableListOf<String>()
-                                                tempCurrentList.add(listCode)
-                                                db.collection("users")
-                                                    .document(Firebase.auth.currentUser!!.uid)
-                                                    .update("lists", tempCurrentList)
-                                            } else {
-                                                val tempCurrentList =
-                                                    document.data?.get("lists") as MutableList<String>
-                                                if (!tempCurrentList.contains(listCode)) {
-                                                    tempCurrentList.add(listCode)
-                                                    db.collection("users")
-                                                        .document(Firebase.auth.currentUser!!.uid)
-                                                        .update("lists", tempCurrentList)
-                                                } else {
-                                                    Toast.makeText(
-                                                        parentActivity,
-                                                        "List already exists",
-                                                        Toast.LENGTH_SHORT
-                                                    ).show()
-                                                }
-                                            }
-                                        }
-                                        val intent = Intent(parentActivity, ScrollListsActivity::class.java)
-                                        startActivity(parentActivity, intent, null)
-                                        parentActivity.finish()
-                                    }
-                    } else {
-                        Toast.makeText(parentActivity, "List not found", Toast.LENGTH_SHORT).show()
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-        /** This is a composable function that creates submenu for creating a new list
-         *
-         * */
-fun CreateListContent(parentActivity: Activity){
-    val colorArray = arrayOf(
-        "Red",
-        "Blue",
-        "Green",
-        "Yellow",
-        "Cyan",
-        "Pink",
-        "White",
-        "Gray",
-    )
-    val db = Firebase.firestore
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Spacer(modifier = Modifier.padding(5.dp))
-        Text(text = "Create a new list", fontSize = 20.sp)
-        Spacer(modifier = Modifier.padding(20.dp))
-        val listName: String = OutlinedTextField("List Name (Required)")
-
-        Spacer(modifier = Modifier.padding(5.dp))
-        val shortDescription: String = OutlinedTextField("Short Description ")
-        Spacer(modifier = Modifier.padding(15.dp))
-        Text(text = "Choose a color for your list", fontSize = 14.sp)
-        val item: String = DropdownMenuBox(colorArray)
-        Spacer(modifier = Modifier.padding(15.dp))
-        HorizontalDivider(modifier = Modifier.height(5.dp))
-        Spacer(modifier = Modifier.padding(10.dp))
-        ButtonFilled("Create") {
-            //TODO: CREATING LIST FUNCTIONALITY
-            Log.d("D", "ListName: $listName")
-            Log.d("D", "ShortDescription: $shortDescription")
-            Log.d("D", "Color: $item")
-            if(listName != ""){
-                val user = User.createUser(Firebase.auth.currentUser?.uid,Firebase.auth.currentUser?.displayName.toString(),Firebase.auth.currentUser?.email.toString())
-                val list = ListOfTasks.createList(listName, user, item, shortDescription)
-                Log.d("D", "User: $user id: ${user.getId()} name: ${user.getName()} email: ${user.getEmail()} lists: ${user.getLists()}")
-                val hashList = hashMapOf(
-                    "listName" to listName,
-                    "code" to list.getCode(),
-                    "color" to item,
-                    "description" to shortDescription,
-                    "creator" to user.getId()
-                )
-                db.collection("lists").document(list.getCode()).set(hashList)
-
-                db.collection("users").document(Firebase.auth.currentUser!!.uid).get()
-                    .addOnSuccessListener { document ->
-                        if (document != null) {
-                            val currentList = document.data?.get("lists")
-
-                            if(currentList == null){
-                                val currentList = mutableListOf<String>()
-                                currentList.add(list.getCode())
-                                db.collection("users").document(Firebase.auth.currentUser!!.uid).update("lists", currentList)
-                            }
-                            else{
-                                val currentList = document.data?.get("lists") as MutableList<String>
-                                currentList.add(list.getCode())
-                                db.collection("users").document(Firebase.auth.currentUser!!.uid).update("lists", currentList)
-                            }
-                            val intent = Intent(parentActivity, ScrollListsActivity::class.java)
-                            startActivity(parentActivity, intent, null)
-                        }
-                    }
-            }
-            else{
-                Toast.makeText(parentActivity, "List name is required", Toast.LENGTH_SHORT).show()
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-        /** This is a composable function that creates a dropdown menu
-         *@param items: Array<String> - The items to be displayed in the dropdown menu
-         *@return selectedText: String - The selected item from the dropdown menu
-         * */
-fun DropdownMenuBox(items: Array<String>): String {
-    val context = LocalContext.current
-    var expanded by remember { mutableStateOf(false) }
-    var selectedText by remember { mutableStateOf(items[0]) }
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(10.dp)
-    ) {
-        ExposedDropdownMenuBox(
-            expanded = expanded,
-            onExpandedChange = {
-                expanded = !expanded
-            }
-        ) {
-            TextField(
-                value = selectedText,
-                onValueChange = {},
-                readOnly = true,
-                trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded) },
-                modifier = Modifier.menuAnchor()
-            )
-
-            ExposedDropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { expanded = false }
-            ) {
-                items.forEach { item ->
-                    DropdownMenuItem(
-                        text = { Text(text = item) },
-                        onClick = {
-                            selectedText = item
-                            expanded = false
-                            Toast.makeText(context, item, Toast.LENGTH_SHORT).show()
-                        }
-                    )
-                }
-            }
-        }
-    }
-    return selectedText
-}
-@Composable
-fun NavDrawer(parentActivity: Activity, listOfLists: List<List<String>>) {
-    ModalDrawerSheet() {
-        Column(
-            modifier = Modifier.fillMaxSize(),
-            verticalArrangement = Arrangement.Bottom,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            // Top section with the text "List it"
-            Spacer(modifier = Modifier.height(100.dp))
-            // Middle section with ListView
-            LazyColumn(
-                modifier = Modifier.weight(1f)
-            ) {
-                items(listOfLists.size) { i ->
-                    NavDrawerItem(
-                        title = listOfLists[i][0],
-                        description = listOfLists[i][1],
-                        color = listOfLists[i][2],
-                        context = parentActivity,
-                        code = listOfLists[i][3]
-                    )
-                }
-            }
-            // Bottom section with buttons
-            HorizontalDivider()
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 10.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                ButtonTonalFilled(label = "Settings") {
-                    //TODO: Implement settings
-                }
-            }
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 10.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                ButtonTonalFilled(label = "Log out") {
-                    //TODO: Implement log out
-                    val intent = Intent(parentActivity, MainActivity::class.java)
-                    startActivity(parentActivity, intent, null)
-                    Firebase.auth.signOut()
-                    parentActivity.finish()
-                }
-            }
-        }
-    }
-}
-
-
-
-@Composable
-fun ListItem(title: String, description: String, color: String, context: Context, code:String) {
-    val backgroundColor = when (color) {
-        "Red" -> Color.Red
-        "Blue" -> Color.Blue
-        "Green" -> Color.Green
-        "Yellow" -> Color.Yellow
-        "Cyan" -> Color.Cyan
-        "Pink" -> Color.Magenta
-        "White" -> Color.White
-        "Gray" -> Color.Gray
-        else -> Color.Transparent
-    }
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(5.dp)
-            .clickable {
-                val intent = Intent(context, ListActivity::class.java)
-                intent.putExtra("listCode", code)
-                context.startActivity(intent)
-            }
-            .background(backgroundColor, shape = RoundedCornerShape(8.dp))
-    ) {
-        Column(
-            modifier = Modifier.padding(horizontal = 10.dp, vertical = 8.dp)
-        ) {
-            Text(
-                text = title,
-                color = Color.Black,
-                fontSize = 25.sp
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = description,
-                color = Color.Black,
-                fontSize = 15.sp
-            )
-        }
-    }
-}
-
-@Composable
-fun NavDrawerItem(title: String, description: String, color: String, context: Context, code:String) {
-    val backgroundColor = Color.Transparent
-    val circleColor = remember {
-        when (color) {
-            "Red" -> Color.Red
-            "Blue" -> Color.Blue
-            "Green" -> Color.Green
-            "Yellow" -> Color.Yellow
-            "Cyan" -> Color.Cyan
-            "Pink" -> Color.Magenta
-            "White" -> Color.White
-            "Gray" -> Color.Gray
-            else -> Color.Transparent
-        }
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(5.dp)
-            .clickable {
-                val intent = Intent(context, ListActivity::class.java)
-                intent.putExtra("listCode", code)
-                context.startActivity(intent)
-            }
-            .background(backgroundColor, shape = RoundedCornerShape(8.dp))
-    ) {
-        Row(modifier = Modifier.padding(horizontal = 10.dp, vertical = 8.dp)) {
-            Box(
-                modifier = Modifier
-                    .size(36.dp)
-                    .background(circleColor, shape = CircleShape)
-                    .align(Alignment.CenterVertically)
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Column {
-                Text(
-                    text = title,
-                    color = Color.Black,
-                    fontSize = 15.sp
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(
-                    text = description,
-                    color = Color.Black,
-                    fontSize = 10.sp
-                )
-            }
-        }
-    }
-}
